refactored transaction_formation to ft_transfer
