+++ src/commands/tokens/send_ft/mod.rs
@@ -1,8 +1,8 @@
 use strum::{EnumDiscriminants, EnumIter, EnumMessage};
 
 mod exact_amount_ft;
+mod ft_transfer;
 mod max_amount_ft;
-mod transaction_formation;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
 #[interactive_clap(input_context = super::TokensCommandsContext)]
